Introduced a `Forecaster` interface
Introduced a `Forecaster` interface, along with my own versions of `Region`, `Day`, and `Forecast`, as well an adapter for the external API. I made copies of `Region`, `Day`, and `Forecast` so that I can own my domain's models.

Right now they are 1:1, but they can very easily diverge in the future if I or the third party add more features. This limits the coupling to the third party API to just the adapter code.

Also introduced a dummy `CacheDecorator` that currently does nothing, and changed `Example` to `Main` and made it use the new `Forecaster` interface API.